brain/pyproject.toml
[project]
name = "brain"
version = "0.1.0"
description = "MCP server: Obsidian vault + server management tools"
requires-python = ">=3.12,<3.14"
dependencies = [
    "mcp[cli]>=1.0",
    "pyyaml>=6.0",
    "psutil>=5.9",
    "faster-whisper>=1.0",
    "httpx>=0.27",
    "tomli>=2.0",
    "onnxruntime>=1.17",
    "tokenizers>=0.15",
    "huggingface-hub>=0.20",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/brain"]
